Move main to separate class
