+++ pom.xml
@@ -10,7 +10,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-        <exec.mainClass>com.vnaskos.pdfguru.PDFGuru</exec.mainClass>
+        <exec.mainClass>com.vnaskos.pdfguru.ui.PDFGuru</exec.mainClass>
     </properties>
 
     <dependencies>
@@ -54,7 +54,7 @@
                     <archive>
                         <manifest>
                             <mainClass>
-                                com.vnaskos.pdfguru.PDFGuru
+                                com.vnaskos.pdfguru.ui.PDFGuru
                             </mainClass>
                         </manifest>
                     </archive>
