+++ src/main/java/Main.java
@@ -1,0 +1,16 @@
+import com.vnaskos.pdfguru.ui.PDFGuru;
+
+import javax.swing.*;
+
+public class Main {
+
+    public static void main(String... args) {
+        try {
+            UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
+        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException | UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(PDFGuru.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+
+        java.awt.EventQueue.invokeLater(() -> new PDFGuru().setVisible(true));
+    }
+}
